wip v2

+++ Engine/Source/Bombast/interfaces.h
@@ -74,7 +74,7 @@ public:
 	virtual void VOnAttach(GameViewId vid, ActorId aid) = 0;
 
 	virtual LRESULT CALLBACK VOnMsgProc(AppMsg msg) = 0;
-	virtual void VOnUpdate(const float deltaMs) = 0;
+	virtual void VOnUpdate(const unsigned long deltaMs) = 0;
 
 	virtual ~IGameView(){};
 };
